Use `api` for marking transitive compile+runtime dependencies
We have previously switched from `compile` to `implementation`.
This had an unexpected consequence, however - `implementation`
does not make dependencies transitive. The quick fix is to
switch back to `compile`, which is marked as deprecated, however.

As an alternative, there exists an `api` label in the `java-library`
Gradle plugin. According to https://stackoverflow.com/a/53726023 ,
this plugin extends the Java plugin and provides functionality
for Java libraries, which is exactly what is needed.