feat: add fast int/string conversion utility
This adds a special class written for fastest possible
conversions compatible with the sysfs filesystem semantics.
This will allow future rewritten Sysfs to be faster. One
example of this is integer parsing - no strings need to be
allocated, they're parsed directly from the ByteBuffer that
was received from read (credits to @dwalend).